+++ src/main/java/OneQ/OnSurvey/domain/survey/model/formRequest/FormConversionPayload.java
@@ -3,6 +3,7 @@ package OneQ.OnSurvey.domain.survey.model.formRequest;
 import java.util.List;
 
 public record FormConversionPayload (
+    Long requestId,
     List<String> urls
 ) {
 }
