+++ src/main/java/OneQ/OnSurvey/domain/survey/service/formRequest/FormCommandService.java
@@ -2,9 +2,11 @@ package OneQ.OnSurvey.domain.survey.service.formRequest;
 
 import OneQ.OnSurvey.domain.member.dto.MemberSearchResult;
 import OneQ.OnSurvey.domain.member.service.MemberFinder;
+import OneQ.OnSurvey.domain.survey.SurveyErrorCode;
 import OneQ.OnSurvey.domain.survey.entity.FormRequest;
 import OneQ.OnSurvey.domain.survey.model.formRequest.FormPublishRequest;
 import OneQ.OnSurvey.domain.survey.model.formRequest.FormValidationAndStashResponse;
+import OneQ.OnSurvey.domain.survey.model.formRequest.FormValidationPayload;
 import OneQ.OnSurvey.domain.survey.model.formRequest.FormValidationRequestDto;
 import OneQ.OnSurvey.domain.survey.model.formRequest.FormValidationResponse;
 import OneQ.OnSurvey.domain.survey.model.formRequest.event.FormRequestConversionEvent;
@@ -15,6 +17,7 @@ import OneQ.OnSurvey.domain.survey.service.command.SurveyCommand;
 import OneQ.OnSurvey.domain.survey.service.query.SurveyQueryService;
 import OneQ.OnSurvey.global.common.exception.CustomException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,6 +26,7 @@ import java.util.List;
 
 import static OneQ.OnSurvey.domain.survey.SurveyErrorCode.*;
 
+@Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
@@ -36,15 +40,18 @@ public class FormCommandService implements FormCreator, FormUpdater, FormPublish
     private final SurveyCommand surveyCommand;
 
     @Override
-    public Long createFormRequest(FormRequestDto dto) {
-        FormRequest request = dto.toEntity();
+    public Long createFormRequest(Long userKey, Long memberId, FormRequestDto dto) {
+        FormRequest request = dto.toEntity(userKey);
         FormRequest savedRequest = formRequestRepository.save(request);
 
         eventPublisher.publishEvent(new FormRequestConversionEvent(
             savedRequest.getId(),
-            savedRequest.getRequesterEmail(),
-            List.of(savedRequest.getFormLink()))
-        );
+            userKey,
+            memberId,
+            List.of(savedRequest.getFormLink()),
+            dto.screening(),
+            dto.surveyForm()
+        ));
         return savedRequest.getId();
     }
 
@@ -95,8 +102,13 @@ public class FormCommandService implements FormCreator, FormUpdater, FormPublish
      */
     @Override
     public FormValidationResponse validationFormRequestLink(FormValidationRequestDto dto) {
-        FormValidationAndStashResponse formCount = formRequestLambda.validateAndStashFormRequest(dto.formLink(), dto.requesterEmail());
+        FormValidationPayload payload = new FormValidationPayload(List.of(dto.formLink()), dto.requesterEmail());
+        FormValidationAndStashResponse validationResult = formRequestLambda.validateAndStashFormRequest(payload);
 
-        return FormValidationResponse.from(formCount);
+        if (validationResult == null || validationResult.successCount() == 0) {
+            log.warn("[FormCommandService:validationFormRequestLink] 구글폼 링크가 유효하지 않음 - URL: {}", dto.formLink());
+            throw new CustomException(SurveyErrorCode.FORM_INVALID);
+        }
+        return FormValidationResponse.from(validationResult);
     }
 }
